fix case.py for bubscreen

+++ tests/working_1d_bubscreen/case.py
@@ -67,7 +67,7 @@ Nt      = int(Tfinal/dt)
 
 Nfiles = 20.
 Nout = int(math.ceil(Nt/Nfiles))
-Nt = Nout*Nfiles
+Nt = int(Nout*Nfiles)
 
 # Command to navigate between directories
 from os import chdir
